feat: add support for KiCad casks
This commit adds support for the KiCad casks, particular the `kicad@8`
cask.

The `kicad` cask is already available from Homebrew core but
version-scoped releases are not available. This is mostly useful when
working between libraries and applications during the period when KiCad
upgrades to the next major version and there is a requirement to open
files using the older release.