docs: amend local dev install info
This commit amends the documented commands to install formulae/casks
from the local working tree.

From Homebrew `4.6.4` onwards, all formulae and casks MUST be installed
from a tap[1], which meant that the existing documented instructions do not
work. For the record, the existing instructions still work when called
with `HOMEBREW_DEVELOPER=1` but it is discouraged[2].

The correct way to do so moving forward is to create a tap locally and
install from the tap.

[1]: https://github.com/Homebrew/brew/pull/20414
[2]: https://github.com/Homebrew/brew/pull/20414#issuecomment-3183734076